refactor: move shared modules into src/shared/ directory
Move functionRegistry, constantEvaluate, nodeTypes, values/, and
matching into src/shared/ so the build config can use directory
boundaries instead of file-by-file matching for chunk assignment.
