+++ src/evaluator/functions/array.ts
@@ -1,4 +1,4 @@
-import {FALSE_VALUE, fromJS, fromString, getType, NULL_VALUE, TRUE_VALUE} from '../../values'
+import {FALSE_VALUE, fromJS, fromString, getType, NULL_VALUE, TRUE_VALUE} from '../../shared/values'
 import {isEqual} from '../equality'
 import {arrayExecutor, mappedExecutor} from '../evaluate'
 import type {FunctionSet} from '.'
