+++ src/evaluator/functions/pt.ts
@@ -1,4 +1,4 @@
-import {fromString, NULL_VALUE} from '../../values'
+import {fromString, NULL_VALUE} from '../../shared/values'
 import {mappedExecutor} from '../evaluate'
 import {portableTextContent} from '../pt'
 import type {FunctionSet} from '.'
