+++ src/evaluator/scoring.ts
@@ -1,5 +1,5 @@
-import type {ExprNode} from '../nodeTypes'
-import type {Value} from '../values'
+import type {ExprNode} from '../shared/nodeTypes'
+import type {Value} from '../shared/values'
 import {executeAsync, executeSync} from './evaluate'
 import {
   type GatheredText,
@@ -7,7 +7,7 @@ import {
   matchPatternRegex,
   matchTokenize,
   type Token,
-} from './matching'
+} from '../shared/matching'
 import {Scope} from './scope'
 
 // BM25 similarity constants
