+++ src/shared/constantEvaluate.ts
