+++ src/shared/functionRegistry.ts
