+++ src/shared/values/StreamValue.ts
