+++ src/traversal.ts
@@ -1,5 +1,5 @@
 /* eslint-disable camelcase */
-import type {ExprNode} from './nodeTypes'
+import type {ExprNode} from './shared/nodeTypes'
 
 export type Traversal = (base: ExprNode) => ExprNode
