+++ test/typeEvaluateCompare.test.ts
@@ -4,7 +4,7 @@ import t from 'tap'
 import {evaluate} from '../src/evaluator'
 import {type GroqFunction, namespaces} from '../src/evaluator/functions'
 import {operators} from '../src/evaluator/operators'
-import type {ExprNode, OpCall} from '../src/nodeTypes'
+import type {ExprNode, OpCall} from '../src/shared/nodeTypes'
 import type {TypeNode} from '../src/typeEvaluator'
 import {DateTime} from '../src/typeEvaluator'
 import {satisfies} from '../src/typeEvaluator/satisfies'
