+++ test/values/StreamValue.test.ts
@@ -1,7 +1,7 @@
 import t from 'tap'
 
-import {fromJS} from '../../src/values'
-import {StreamValue} from '../../src/values/StreamValue'
+import {fromJS} from '../../src/shared/values'
+import {StreamValue} from '../../src/shared/values/StreamValue'
 
 t.test('StreamValue', async (t) => {
   t.test('handles exceptions during fetch', async (t) => {
